Reconsider writing this in the quirky ABOP syntax
The parser for the language used in ABOP is getting really complicated.
I will ultimately be parsing all the expressions into Haskell. I expect
the users to mostly be from the Haskell world (well, I may be the only
user). It may be easier to implement this as a domain specific language.