Implement stochastic rules
My random number handling seems really awkward. Currently I am passing a
random number generator all over the program, splitting it as a recurse.
It is bogging all my functions down in boilerplate. And for all this
mud, I use the generator only in one place, the choose function.

But, however ugly it is, it is working now. So the next thing I will do
is write some sugar to allow easy creation of simple stochastic models.